Rename the NAA file
