+++ packages/teams-js/src/internal/nestedAppAuth.ts
